services/acquisition/catalog-collector/pyproject.toml
[project]
name = "catalog-collector"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.2"
dependencies = [
    "aiohttp>=3.13.2",
    "aiokafka>=0.12.0",
    "alembic>=1.17.2",
    "apscheduler>=3.11.2",
    "asyncpg>=0.30.0",
    "backports-zstd>=1.2.0",
    "bcrypt>=4.3.0",
    "beautifulsoup4>=4.14.3",
    "brotli>=1.2.0",
    "coloredlogs>=15.0.1",
    "colorlog>=6.10.1",
    "fastapi>=0.127.0",
    "html5lib>=1.1",
    "icecream>=2.1.8",
    "lxml>=5.4.0",
    "playwright>=1.57.0",
    "psycopg>=3.3.2",
    "psycopg-binary>=3.3.2",
    "psycopg2>=2.9.11",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.1",
    "pytz>=2025.2",
    "requests>=2.32.5",
    "requests-oauthlib>=2.0.0",
    "slugify>=0.0.1",
    "sqlalchemy>=2.0.45",
    "starlette>=0.46.2",
    "tabulate>=0.9.0",
    "tqdm>=4.67.1",
    "uvicorn>=0.34.3",
    "watchfiles>=1.1.1",
]

[tool.poetry]
package-mode = false


[tool.pytest.ini_options]
pythonpath = [
    "src"
]
